Master mit halbwegs funktionierendem Programm ersetzt

+++ pages-pwreset.php
@@ -6,23 +6,16 @@ if(session_status() === PHP_SESSION_NONE)
 {
     session_start();
 }
-include 'config.php';
+include_once 'functions.php';
 ?>
 <?php
-include "head.php";
+include_once "head.php";
 ?>
 <?php
-include "header.php";
+include_once "header.php";
 ?>
 <?php
-include "Sidebar.php";
-?>
-
-<?php
-/*if (isset($_POST['ResetCheck'])){
-    include "pages-pwreset2.php";
-}*/
-//if (!isset($_POST['ResetConfirmation'])){
+include_once "Sidebar.php";
 ?>
 
 <div class="background-image"></div>
@@ -63,13 +56,6 @@ include "Sidebar.php";
                                     </div>
 
                                     <div class=" mb-2 text-center">
-                                        <?php
-                                        if (isset($_POST['ResetCheck'])){
-                                            echo '<input type="hidden" name="email1" value="'.$_POST['emailReset'].'">';
-                                            echo '<input type="hidden" name="username1" value="'.$_POST['usernameReset'].'">';
-                                        }
-
-                                        ?>
                                         <input name="ResetConfirmation" class="btn btn-danger btn-lg btn-block" type="submit" value="Zurücksetzen">
                                     </div>
                                     <hr>
@@ -90,58 +76,42 @@ include "Sidebar.php";
                                         <div class="d-flex justify-content-between">
                                             <u><a href="contact.php" class="small  registertxt">Hilfe</a></u>
                                             <u><a href="about.php" class="small  registertxt">Fragen</a></u>
-                                            <button class="btn btn-danger btn-md" onclick="history.back(-2)">Zurück</button>
+                                            <button class="btn btn-danger btn-md" onclick="history.back()">Zurück</button>
                                         </div>
                                     </div>
                                     <?php
                                         global $con;
-                                    /*} else */
-                                    if (isset($_POST['ResetConfirmation'])){
+                                        if (isset($_POST['ResetConfirmation'])){
                                             try
                                             {
-                                                $input_username = "";
-                                                $input_email = "";
-                                                $input_newpw = "";
-                                                $input_newpwconfirm = "";
-                                                //if(isset())
-                                                //if (isset($_POST['ResetCheck']))
-                                                //{
-                                                $input_username = $_POST["username1"];
-                                                $input_email = $_POST["email1"];
-                                                //}
-
-                                                //$input_username = $_SESSION['usernameResetSession'];
-                                                //$input_email = $_SESSION['emailResetSession'];
-                                                //echo '<script>alert("Username:'.$_POST["username1"].'  '.print_r($_POST).' Email:'.$_POST['email1'].' PW:'.$input_newpw.'")</script>';
-                                                //echo ''.$input_username.' '.$input_email;
+                                                $input_username = $_POST["usernameReset"];
+                                                $input_email = $_POST["emailReset"];
+                                                $input_newpw = $_POST["passwordReset"];
+                                                $input_newpwconfirm = $_POST["passwordResetConfirm"];
 
                                                 $idOfUser = 0;
 
-                                                $querySelectid = 'Select user_id from user WHERE username = ? AND email = ?';
-                                                $sqlSelectid = $con->prepare($querySelectid);
-                                                $sqlSelectid->execute([trim($input_username), trim($input_email)]);
-
-                                                $rc = $sqlSelectid->rowCount();
-                                                echo '<script>alert("'.$input_username.' '.$input_email.' '.$rc.'")</script>';
-
+                                                $querySelectid = 'Select user_id from skimp.user WHERE username = ? AND email = ?';
+                                                $sqlSelectid = $_SESSION['DBConnection']->prepare($querySelectid);
+                                                $sqlSelectid->execute([$input_username, $input_email]);
 
                                                 while($row1 = $sqlSelectid->fetch(PDO::FETCH_NUM))
                                                 {
                                                     $idOfUser = $row1[0];
                                                 }
+                                                $rc = $sqlSelectid->rowCount();
 
-                                                $input_newpw = $_POST['passwordReset'];
-                                                $input_newpwconfirm = $_POST['passwordResetConfirm'];
+                                                $input_newpw = trim($_POST['passwordReset']);
+                                                $input_newpwconfirm = trim($_POST['passwordResetConfirm']);
 
-                                                if($input_newpw == $input_newpwconfirm && $rc > 0)
+                                                if($input_newpw == $input_newpwconfirm && $rc > 0 && !empty($input_newpw) && strlen($input_newpw) >= 6)
                                                 {
                                                     try
                                                     {
-                                                        $queryPWReset = 'UPDATE skimp.user SET password = ? WHERE user_id = ?';
+                                                        $queryPWReset = 'UPDATE user SET password = ? WHERE user_id = ?';
                                                         $sqlPWReset = $con->prepare($queryPWReset);
                                                         $sqlPWReset->execute([md5($input_newpw), $idOfUser]);
 
-                                                        //unset($_POST['ResetConfirmation']);
                                                         $alertChangedPW = "Passwort erfolgreich verändert";
                                                         //echo "<a href=index.php>$alert</a>";
                                                         echo '<script>alert("'.$alertChangedPW.'")</script>';
@@ -162,12 +132,20 @@ include "Sidebar.php";
                                                     };*/
                                                     $alert = "Inkorrekter (oder nicht zusammengehörender) Username und/oder Passwort";
                                                     echo '<script>alert("'.$alert.'")</script>';
+                                                    //echo '<a href="page1.php">';
                                                     //header("Location: http://localhost:63342/Projektpraktikum/index.php");
                                                     //echo '<button onclick="history.back(-2)">Zurück!</button>';
                                                 }
                                                 elseif ($input_newpw != $input_newpwconfirm)
                                                 {
                                                     $alert = "Passwörter stimmen nicht überein";
+                                                    //echo "<a href=index.php>$alert</a>";
+                                                    echo '<script>alert("'.$alert.'")</script>';
+                                                    //echo '<button onclick="history.back(-2)">Zurück!</button>';
+                                                }
+                                                elseif (strlen($input_newpw) < 6)
+                                                {
+                                                    $alert = "Passworter zu kurz";
                                                     //echo "<a href=index.php>$alert</a>";
                                                     echo '<script>alert("'.$alert.'")</script>';
                                                     //echo '<button onclick="history.back(-2)">Zurück!</button>';
@@ -198,7 +176,7 @@ include "Sidebar.php";
     </div>
 </section>
 <?php
-include "footer.php";
+include_once "footer.php";
 ?>
 </body>
